+++ packages/build-utils/src/index.ts
@@ -57,7 +57,10 @@ import debug from './debug';
 import getIgnoreFilter from './get-ignore-filter';
 import { getPlatformEnv } from './get-platform-env';
 import { getPrefixedEnvVars } from './get-prefixed-env-vars';
-import { getServiceUrlEnvVars } from './get-service-url-env-vars';
+import {
+  getServiceUrlEnvVars,
+  getExperimentalServiceUrlEnvVars,
+} from './get-service-url-env-vars';
 import { cloneEnv } from './clone-env';
 import { hardLinkDir } from './hard-link-dir';
 import { validateNpmrc } from './validate-npmrc';
@@ -109,6 +112,7 @@ export {
   getPlatformEnv,
   getPrefixedEnvVars,
   getServiceUrlEnvVars,
+  getExperimentalServiceUrlEnvVars,
   streamToBuffer,
   streamToBufferChunks,
   debug,
@@ -133,6 +137,7 @@ export { getOsRelease, getProvidedRuntime } from './os';
 export * from './should-serve';
 export * from './schemas';
 export * from './package-manifest';
+export { generateProjectManifest } from './node-diagnostics';
 export * from './types';
 export * from './errors';